Fix dlib CI issue
Add dlib to setup.py

Add pip cache and increase timeout

Change timeout to 40 min

Switch to just 3.8 testing

Restructure ml module

Remove timeout and add 3.9 as testing option

Change import in ml test

Update CI
