+++ .github/workflows/build.yml
@@ -16,9 +16,19 @@ jobs:
         with:
           python-version: ${{ matrix.python }}
 
-      - name: Install dependencies and script
+      - uses: actions/cache@v2
+        with:
+          path: ~/.cache/pip
+          key: ${{ runner.os }}-pip-${{ hashFiles('**/requirements.txt') }}
+          restore-keys: |
+            ${{ runner.os }}-pip-
+
+      - name: Install test dependencies and script
         run: pip install .[tests]
       
+      - name: Install dependencies and script
+        run: pip install -r requirements.txt
+
       - name: Lint with flake8
         run: flake8
