+++ src/openvtuber/ml/__init__.py
@@ -1,4 +1,5 @@
-from .ml import infer, infer_image
+from .ml import infer
+from .face import infer_image
 
 
 __all__ = ['infer', 'infer_image', ]
