+++ tests/ml_test.py
@@ -1,5 +1,5 @@
 import pytest
-from openvtuber import ml
+from openvtuber.ml import infer
 import cv2
 import os
 
@@ -9,5 +9,5 @@ def test_infer(capsys):
     path = os.path.dirname(__file__)
     rel_path = 'vscout.png'
     in_img = cv2.imread(os.path.join(path, rel_path), cv2.IMREAD_COLOR)
-    img = ml.infer(in_img)
+    img = infer(in_img)
     assert (img.size != 0)
